backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "timetracker42-backend"
version = "0.1.0"
description = "Backend service for TimeTracker42"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.35,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "alembic>=1.14.0,<2.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.7.0,<1.0.0",
  "mypy>=1.13.0,<2.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["I001"]
"alembic/versions/*.py" = ["E501", "I001"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*"]
